Fix byparts keyword argument
Now magnetization function actually allows for selection of method to use for integration.

Did a variety of tests (in the notebook with changes in this commit). Both integrals agree for the most part, but there may be slight discrepancies for small strain, Fermi energy near the band edge, and/or low temperature.