Start rewriting code with full tight binding model
Can calculate full tight binding model for arbitrary K. This can either be over the full Brillouin zone or in a small region near a Dirac point. The application of strain shifts the high-symmetry points of the Brillouin zone, but because the bands depend on the change in geometry as well as a change in the hopping parameters, the Dirac points shift to a different location. This location is difficult to write down accurately, even in the limit of small strain.