Fix carrier density and zero energy definitions
Zero energy defined as the energy at which the net carrier density is zero at zero temperature. The carrier density is defined in terms of hole occupation in the two lowest-energy bands. This is so the Fermi occupation goes to zero at the edges of the window of k-space we consider.