Clarify negative sign in Hamiltonian derivative
I use the same bond vectors delta for all types of hopping, but the directions of the bond vectors are opposite for gamma3 hopping. I simply included a negative sign when constructing the Hamiltonian. However, it is important to note that this negative sign does NOT appear in front of the derivative (i.e. via the chain rule).